Updates from: https://github.com/malucooo/Otxserver-New

+++ path_10_9/src/gamestore.cpp
@@ -17,6 +17,8 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
+#include "otpch.h"
+
 #include "gamestore.h"
 
 #include "pugicast.h"
@@ -308,7 +310,7 @@ HistoryStoreOfferList IOGameStore::getHistoryEntries(uint32_t account_id, uint32
 
     query << "SELECT `description`,`mode`,`coin_amount`,`time` FROM `store_history` WHERE `account_id` = " <<account_id << " ORDER BY `time` DESC LIMIT "
           << (std::max<int>((page-1),0)*GameStore::HISTORY_ENTRIES_PER_PAGE)
-          << "," << (uint)GameStore::HISTORY_ENTRIES_PER_PAGE <<";";
+          << "," << (uint16_t)GameStore::HISTORY_ENTRIES_PER_PAGE <<";";
     DBResult_ptr result = Database::getInstance().storeQuery(query.str());
 
     if(result) {
